# src/api/openapi.yml
openapi: 3.0.0
info:
  version: 0.1.0
  title: AgentENV API
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    # AuthProviderBearerAuth / AuthProviderTeamAuth: B before T in the name
    # so Bearer is validated before Team.
    AuthProviderBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: access_token
    AuthProviderTeamAuth:
      type: apiKey
      in: header
      name: X-Team-ID
    # AdminApiKeyAuth / AdminTeamAuth: alphabetical names ensure token
    # validation runs before team context population.
    AdminApiKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Token
    AdminTeamAuth:
      type: apiKey
      in: header
      name: X-Team-ID
  parameters:
...
    sandboxID:
      name: sandboxID
      in: path
      required: true
      schema:
        type: string
...
    paginationLimit:
      name: limit
      in: query
      description: Maximum number of items to return per page
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        default: 100
        maximum: 100
    paginationNextToken:
      name: nextToken
      in: query
      description: Cursor to start the list from
      required: false
      schema:
        type: string
  headers:
    XNextToken:
      description: Cursor to fetch the next page of results, if more exist
      schema:
        type: string
    XTotalRunning:
      description: >
        Number of running sandboxes matching the filters, before pagination is applied.
        Only present when running sandboxes were requested.
      schema:
        type: integer
        format: int32
  responses:
    "400":
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    "401":
      description: Authentication error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    "404":
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    "409":
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    "500":
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    SandboxMetadata:
      additionalProperties:
        type: string
        description: Metadata of the sandbox
    SandboxState:
      type: string
      description: State of the sandbox
      enum:
        - running
        - paused
        - cleanup_pending
    OrderDirection:
      type: string
      description: Sort direction
      default: desc
      enum:
        - asc
        - desc
...
    EnvVars:
      additionalProperties:
        type: string
        description: Environment variables for the sandbox
...
    Mcp:
      type: object
      description: MCP configuration for the sandbox
      additionalProperties: {}
      nullable: true
    CustomExtensionParams:
      type: object
      additionalProperties: true
      description: >
        Opaque JSON object interpreted only by the custom extension. An
        absent value and an empty object are equivalent: both mean empty
        params.
    SandboxNetworkConfig:
      type: object
      properties:
        allowPublicTraffic:
          type: boolean
          default: true
          description: Specify if the sandbox URLs should be accessible only with authentication.
        allowOut:
          type: array
          description: List of allowed destinations for egress traffic. Each entry can be a CIDR block (e.g. "8.8.8.8/32"), a bare IP address (e.g. "8.8.8.8"), or a domain name (e.g. "example.com", "*.example.com"). Allowed entries always take precedence over denied entries.
          items:
            type: string
        denyOut:
          type: array
          description: List of denied CIDR blocks or IP addresses for egress traffic. Domain names are not supported for deny rules.
          items:
            type: string
        maskRequestHost:
          type: string
          description: Specify host mask which will be used for all sandbox requests
...
    SandboxAutoResumeConfig:
      type: object
      description: Auto-resume configuration for paused sandboxes.
      required:
        - enabled
      properties:
        enabled:
          $ref: "#/components/schemas/SandboxAutoResumeEnabled"
...
    Sandbox:
      required:
        - templateID
        - sandboxID
        - clientID
        - envdVersion
      properties:
        templateID:
          type: string
          description: Identifier of the template from which is the sandbox created
        sandboxID:
          type: string
          description: Identifier of the sandbox
        alias:
          type: string
          description: Alias of the template
        clientID:
          type: string
          deprecated: true
          description: Identifier of the client
        envdVersion:
          $ref: "#/components/schemas/EnvdVersion"
        envdAccessToken:
          type: string
          description: Access token used for envd communication
        trafficAccessToken:
          type: string
          nullable: true
          description: Token required for accessing sandbox via proxy.
        domain:
          type: string
          nullable: true
          description: Base domain where the sandbox traffic is accessible
    SandboxDetail:
      required:
        - templateID
        - sandboxID
        - clientID
        - startedAt
        - cpuCount
        - memoryMB
        - diskSizeMB
        - endAt
        - state
        - envdVersion
      properties:
        templateID:
          type: string
          description: Identifier of the template from which is the sandbox created
        alias:
          type: string
          description: Alias of the template
        sandboxID:
          type: string
          description: Identifier of the sandbox
        clientID:
          type: string
          deprecated: true
          description: Identifier of the client
        startedAt:
          type: string
          format: date-time
          description: Time when the sandbox was started
        endAt:
          type: string
          format: date-time
          description: Time when the sandbox will expire
        envdVersion:
          $ref: "#/components/schemas/EnvdVersion"
        envdAccessToken:
          type: string
          description: Access token used for envd communication
        allowInternetAccess:
          type: boolean
          nullable: true
          description: Whether internet access was explicitly enabled or disabled for the sandbox. Null means it was not explicitly set.
        domain:
          type: string
          nullable: true
          description: Base domain where the sandbox traffic is accessible
        cpuCount:
          $ref: "#/components/schemas/CPUCount"
        memoryMB:
          $ref: "#/components/schemas/MemoryMB"
        diskSizeMB:
          $ref: "#/components/schemas/DiskSizeMB"
        metadata:
          $ref: "#/components/schemas/SandboxMetadata"
        state:
          $ref: "#/components/schemas/SandboxState"
        network:
          $ref: "#/components/schemas/SandboxNetworkConfig"
        lifecycle:
          $ref: "#/components/schemas/SandboxLifecycle"
        volumeMounts:
          $ref: "#/components/schemas/VolumeMounts"
    ListedSandbox:
      required:
        - templateID
        - sandboxID
        - clientID
        - startedAt
        - cpuCount
        - memoryMB
        - diskSizeMB
        - endAt
        - state
        - envdVersion
      properties:
        templateID:
          type: string
          description: Identifier of the template from which is the sandbox created
        alias:
          type: string
          description: Alias of the template
        sandboxID:
          type: string
          description: Identifier of the sandbox
        clientID:
          type: string
          deprecated: true
          description: Identifier of the client
        startedAt:
          type: string
          format: date-time
          description: Time when the sandbox was started
        endAt:
          type: string
          format: date-time
          description: Time when the sandbox will expire
        cpuCount:
          $ref: "#/components/schemas/CPUCount"
        memoryMB:
          $ref: "#/components/schemas/MemoryMB"
        diskSizeMB:
          $ref: "#/components/schemas/DiskSizeMB"
        metadata:
          $ref: "#/components/schemas/SandboxMetadata"
        state:
          $ref: "#/components/schemas/SandboxState"
        envdVersion:
          $ref: "#/components/schemas/EnvdVersion"
        volumeMounts:
          $ref: "#/components/schemas/VolumeMounts"
    SandboxAllocationOwner:
      type: object
      required: [version, ownerId]
      properties:
        version:
          type: integer
          enum: [1]
        ownerId:
          type: string
    SandboxAllocation:
      type: object
      required: [version, ownerId, allocationId, sandboxId, requestDigest, state, cancelRequested]
      properties:
        version:
          type: integer
          enum: [1]
        ownerId:
          type: string
        allocationId:
          type: string
        sandboxId:
          type: string
          nullable: true
        requestDigest:
          type: string
          nullable: true
        state:
          type: string
          enum: [pending, settled, cancelled, interrupted]
          description: Settled means the create handler ended, not that the guest was deleted. Interrupted requires host reconciliation; an empty in-memory inventory is insufficient.
        cancelRequested:
          type: boolean
          description: Prevents a later request from claiming this key. An already pending operation is allowed to settle before guest cleanup.
    NewSandbox:
      required:
        - templateID
      properties:
        templateID:
          type: string
          description: >
            Identifier of the required template/snapshot. Use
            POST /sandboxes-cold to create a sandbox directly from an external
            OCI image.
        timeout:
          type: integer
          format: int32
          minimum: 0
          default: 15
          description: Time to live for the sandbox in seconds.
        autoPause:
          type: boolean
          default: true
          description: Automatically pauses the sandbox after the timeout
        autoResume:
          $ref: "#/components/schemas/SandboxAutoResumeConfig"
        secure:
          type: boolean
          description: Secure all system communication with sandbox
        allow_internet_access:
          type: boolean
          description:
            Allow sandbox to access the internet. When set to false, it behaves the same as specifying denyOut
            to 0.0.0.0/0 in the network config.
        network:
          $ref: "#/components/schemas/SandboxNetworkConfig"
        metadata:
          $ref: "#/components/schemas/SandboxMetadata"
        envVars:
          $ref: "#/components/schemas/EnvVars"
        customExtensionParams:
          $ref: "#/components/schemas/CustomExtensionParams"
        mcp:
          $ref: "#/components/schemas/Mcp"
        volumeMounts:
          $ref: "#/components/schemas/VolumeMounts"
...
    ConnectSandbox:
      type: object
      required:
        - timeout
      properties:
        timeout:
          description: Timeout in seconds from the current time after which the sandbox should expire
          type: integer
          format: int32
          minimum: 0
...
    Error:
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: Error code
        message:
          type: string
          description: Error
...
paths:
...
  /sandbox-allocation-owner:
    get:
      summary: Read the durable allocation state owner
      description: Persist this owner ID in the client intent before sending a keyed create. Supply it on all allocation requests; a different node or recreated database must not settle the original work.
      tags: [sandboxes]
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Allocation owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SandboxAllocationOwner"
        "401":
          $ref: "#/components/responses/401"
        "500":
          $ref: "#/components/responses/500"
  /sandbox-allocations/{allocationID}:
    parameters:
      - name: X-Agentenv-Allocation-Owner
        in: header
        required: true
        description: Durable owner ID read before the original create intent. A mismatch returns 409 without claiming or cancelling the key.
        schema:
          type: string
      - name: allocationID
        in: path
        required: true
        schema:
          type: string
    post:
      summary: Create a sandbox with durable allocation ownership
      description: >
        Commit a server-assigned sandbox ID before template resolution, volume
        restoration or VM launch. The allocationID is a non-nil UUID and is
        never reusable. Repeated requests return 409 without new work; changed
        request bodies are rejected. Query the receipt after a lost response.
        This dedicated route fails on servers without allocation support.
      tags: [sandboxes]
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewSandbox"
      responses:
        "201":
          description: The sandbox was created successfully
          headers:
            x-agentenv-sandbox-id:
              description: Identifier of the created sandbox for proxy routing.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sandbox"
        "401":
          $ref: "#/components/responses/401"
        "400":
          $ref: "#/components/responses/400"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"

    get:
      summary: Read the original keyed create outcome
      description: Read-only lookup. A 404 is absence of a receipt, not proof that a delayed POST cannot still arrive. Keys and receipts are node-local and must be routed to the original durable state owner.
      tags: [sandboxes]
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Allocation receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SandboxAllocation"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"
    delete:
      summary: Fence a keyed create operation
      description: Durably prevents a not-yet-arrived request from creating a guest. Does not delete a guest. If a create is pending, wait for settlement and then clean up the original sandbox. Interrupted receipts require host reconciliation. Retain cancellation tombstones across restarts.
      tags: [sandboxes]
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Allocation receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SandboxAllocation"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"
  /sandboxes:
    get:
      summary: List running sandboxes
      description: List all running sandboxes. Use GET /v2/sandboxes instead.
      deprecated: true
      tags: [sandboxes]
      security:
        - ApiKeyAuth: []
        - AuthProviderBearerAuth: []
          AuthProviderTeamAuth: []
        - AdminApiKeyAuth: []
          AdminTeamAuth: []
      parameters:
        - name: metadata
          in: query
          description: Metadata query used to filter the sandboxes (e.g. "user=abc&app=prod"). Each key and values must be URL encoded.
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Successfully returned all running sandboxes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ListedSandbox"
        "401":
          $ref: "#/components/responses/401"
        "400":
          $ref: "#/components/responses/400"
        "500":
          $ref: "#/components/responses/500"
    post:
      summary: Create sandbox
      description: >
        Create a sandbox from an existing template/snapshot. Use
        POST /sandboxes-cold to create a sandbox directly from an external OCI
        image.
      tags: [sandboxes]
      security:
        - ApiKeyAuth: []
        - AuthProviderBearerAuth: []
          AuthProviderTeamAuth: []
        - AdminApiKeyAuth: []
          AdminTeamAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewSandbox"
      responses:
        "201":
          description: The sandbox was created successfully
          headers:
            x-agentenv-sandbox-id:
              description: Identifier of the created sandbox for proxy routing.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sandbox"
        "401":
          $ref: "#/components/responses/401"
        "400":
          $ref: "#/components/responses/400"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"
...
  /v2/sandboxes:
    get:
      summary: List sandboxes (v2)
      description: List all sandboxes
      tags: [sandboxes]
      security:
        - ApiKeyAuth: []
        - AuthProviderBearerAuth: []
          AuthProviderTeamAuth: []
        - AdminApiKeyAuth: []
          AdminTeamAuth: []
      parameters:
        - name: metadata
          in: query
          description: Metadata query used to filter the sandboxes (e.g. "user=abc&app=prod"). Each key and values must be URL encoded.
          required: false
          schema:
            type: string
        - name: state
          in: query
          description: Filter sandboxes by one or more states
          required: false
          schema:
            type: array
            items:
              $ref: "#/components/schemas/SandboxState"
          style: form
          explode: false
        - name: order
          in: query
          description: Sort direction by sandbox start time. Defaults to desc (newest first).
          required: false
          schema:
            $ref: "#/components/schemas/OrderDirection"
        - name: startedAfter
          in: query
          description: Return sandboxes started at or after this timestamp.
          required: false
          schema:
            type: string
            format: date-time
        - name: template
          in: query
          description: Filter sandboxes by a template ID or alias.
          required: false
          schema:
            type: string
        - $ref: "#/components/parameters/paginationNextToken"
        - $ref: "#/components/parameters/paginationLimit"
      responses:
        "200":
          description: Successfully returned all running sandboxes
          headers:
            X-Next-Token:
              $ref: "#/components/headers/XNextToken"
            X-Total-Running:
              $ref: "#/components/headers/XTotalRunning"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ListedSandbox"
        "401":
          $ref: "#/components/responses/401"
        "400":
          $ref: "#/components/responses/400"
        "500":
          $ref: "#/components/responses/500"
  /sandboxes/{sandboxID}:
    get:
      summary: Sandbox
      description: Get a sandbox by id
      tags: [sandboxes]
      security:
        - ApiKeyAuth: []
        - AuthProviderBearerAuth: []
          AuthProviderTeamAuth: []
        - AdminApiKeyAuth: []
          AdminTeamAuth: []
      parameters:
        - $ref: "#/components/parameters/sandboxID"
      responses:
        "200":
          description: Successfully returned the sandbox
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SandboxDetail"
        "404":
          $ref: "#/components/responses/404"
        "401":
          $ref: "#/components/responses/401"
        "500":
          $ref: "#/components/responses/500"

    delete:
      summary: Kill sandbox
      description: Kill a sandbox
      tags: [sandboxes]
      security:
        - ApiKeyAuth: []
        - AuthProviderBearerAuth: []
          AuthProviderTeamAuth: []
        - AdminApiKeyAuth: []
          AdminTeamAuth: []
      parameters:
        - $ref: "#/components/parameters/sandboxID"
      responses:
        "204":
          description: The sandbox was killed successfully
        "404":
          $ref: "#/components/responses/404"
        "401":
          $ref: "#/components/responses/401"
        "500":
          $ref: "#/components/responses/500"

  # TODO: Pause and resume might be exposed as POST /sandboxes/{sandboxID}/snapshot and then POST /sandboxes with specified snapshotting setup
...
  /sandboxes/{sandboxID}/connect:
    post:
      summary: Connect sandbox
      description: Returns sandbox details. If the sandbox is paused, it will be resumed. TTL is only extended.
      tags: [sandboxes]
      security:
        - ApiKeyAuth: []
        - AuthProviderBearerAuth: []
          AuthProviderTeamAuth: []
        - AdminApiKeyAuth: []
          AdminTeamAuth: []
      parameters:
        - $ref: "#/components/parameters/sandboxID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectSandbox"
      responses:
        "200":
          description: The sandbox was already running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sandbox"
        "201":
          description: The sandbox was resumed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sandbox"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"
...